feat:create search feature
Improve styles to make every card content to show similar regardless the
image size. Refactor functions to improve legibility and mantenibility
by abstracting the code and placing it into helpers.js. Lastly, create
another 'print' helper function to render a card list for the results
and use it into the search module when adding the event listener to the
search subscriber